refactor(preview-actions): extract access request and cache helpers
